# src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Vigilancia - Porterias APP API.",
    "version": "1.0.1",
    "description": "Documentación de Vigilancias - Porterias API"
  },
  "paths": {
    "/api/v1/accounts/create": {
      "post": {
        "summary": "Crea una nueva cuenta",
        "tags": [
          "Accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta creada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "error fetching data"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/update": {
      "put": {
        "summary": "Actualiza una cuenta existente",
        "tags": [
          "Accounts"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAccountRequestModel"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Cuenta actualizada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponseModel"
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/delete/{account_id}": {
      "delete": {
        "summary": "Elimina una cuenta",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "account_id",
            "required": true,
            "schema": {
              "type": "integer"
            },
            "description": "ID de la cuenta a eliminar"
          }
        ],
        "responses": {
          "200": {
            "description": "Cuenta eliminada exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/accounts/all": {
      "get": {
        "summary": "Obtiene todas las cuentas",
        "tags": [
          "Accounts"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer"
            },
            "description": "Número de página para la paginación"
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer"
            },
            "description": "Número de elementos por página"
          },
          {
            "in": "query",
            "name": "sortBy",
            "schema": {
              "type": "string"
            },
            "description": "Campo por el cual ordenar los resultados"
          },
          {
            "in": "query",
            "name": "sortOrder",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            },
            "description": "Orden de clasificación (ascendente o descendente)"
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de cuentas obtenida exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "page": {
                      "type": "string"
                    },
                    "pageSize": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error del servidor",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "AccountRequestModel": {
        "type": "object",
        "required": [
          "name",
          "domain"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Nombre de la cuenta"
          },
          "domain": {
            "type": "string",
            "description": "Dominio asociado a la cuenta"
          },
          "parent_account_id": {
            "type": "integer",
            "description": "ID de la cuenta padre (opcional)"
          }
        }
      },
      "UpdateAccountRequestModel": {
        "type": "object",
        "required": [
          "id"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID de la cuenta a actualizar"
          },
          "name": {
            "type": "string",
            "description": "Nuevo nombre de la cuenta (opcional)"
          },
          "domain": {
            "type": "string",
            "description": "Nuevo dominio de la cuenta (opcional)"
          },
          "parent_account_id": {
            "type": "integer",
            "description": "Nuevo ID de la cuenta padre (opcional)"
          }
        }
      },
      "AccountResponseModel": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "ID único de la cuenta"
          },
          "name": {
            "type": "string",
            "description": "Nombre de la cuenta"
          },
          "domain": {
            "type": "string",
            "description": "Dominio asociado a la cuenta"
          },
          "parent_account_id": {
            "type": "integer",
            "description": "ID de la cuenta padre (opcional)"
          }
        }
      },
...
    }
  },
...
}